Move helper methods into convert.launcher module.
